requirements.txt
numpy==1.21.5
cloudpickle>=1.2.0
importlib_metadata>=4.8.0
dataclasses
typing_extensions==4.3.0
lz4>=3.1.0
matplotlib==3.6.2
keyboard==0.13.5
box2d-py==2.3.5
mujoco_py<2.2,>=2.1
gymnasium==0.27.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.0+cu116
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.14.0+cu116
ray==2.0.1
pygame